Marking comments as outdated must not be guarded
Marking comments as outdated does have nothing to do with editing
the comment. This will be done, whenever a new commit changes the
file of the comment (if this is not a global one).

Such commits require the 'push' permission. Having such a permission
does not imply, that the user does have the 'modify comments'
permission, too.

Besides this code is called from a hook only, not via rest endpoints.
This is only internal logic.