Add reply event and simplify internal api
To prepare events for replies we add explicit events for them.
To send mails for replies, we need to have the ability to send mails
independent of the subscriber/reviewer logic. Therefore we pull the
computation of the receiver logic up in the process, where the hooks
are evaluated. Doing so we can remove most distinctions between
subscribers and reviewers, because only the pr events (create, modify
and delete) need this distinction.