Extend CI permission to read pull requests
To enable a jenkins with the scm-manager plugin to find
pull requests when the user configured in jenkins only has
the permission role 'CI-SERVER' we have to extend this
role to read pull requests, too. This has the downside,
that we will introduce this role if the CI plugin is not
installed, but we prefer this over a jenkins instance not
able to read pull requests.