Do not check hooks for merges on internal merge
We do not want to run the merge detection and check hooks
when we do a merge using the merge service. Otherwise we
would do checks multiple times, fire events multiple times
and create comments multiple times.

To prevent all this we reuse the "runPrivileged" mechanism
from the BranchProtectionHook for the MergeObstacleCheckHook
and the StatusCheckHook.

While we are at it we remove the override message parameter
from the "normal" merge service call, because a normal merge
does not have an override message.