Detect deleted branches without merge detection
When a branch is deleted, corresponding pull requests should be
rejected even when the event does not support merge detection,
because this feature is irrelevant for this.

Due to this bug deleted branches did not lead to the rejection of
pull requests when the branch was deleted from the gui.