Fix outdated tag without context being displayed as clickable (#125)
When a pull request with a comment on it changes, the comment is displayed as outdated. Comments which referred to a line in the source code should have a clickable tag on them to open the original information. Comments which did not refer to a line in code, should not but were incorrectly displayed as such. All that needed to be done is set the default click handler to undefined by default and mark all click handlers as optional. The tag already handled the correct highlighting based on the presence of the click handler.