Fix pagination with status filters (#193)
Fix pagination with status filter. The status filter is now being controlled by the query param to ensure the state is still the same after the page was reloaded.

Before the status had been resetted on page change and the page number did stay the same after changing the filter, probably leading to an error if the page does not exist for the new filter.