Fix branch name in PR tags (#208)
The branch names of branches involved in a pull request may be long and not fit the space in the ui. To make sure, that they are easily recognized the first part of the name will always be visible. A title-tag ensures, that the full name is accessible. The clipped part is marked with ellipsis.

Co-authored-by: Konstantin Schaper <konstantin.schaper@cloudogu.com>