Add additional status information to pr table (#215)
This pr adds two things. First, it adds a new column to the pr table which displays the status of the workflow engine. If the engine is not enabled either globally or in the repository or if there are no rules configured, the column is hidden. Second, it adds a new extension point to the pr table that allows adding further columns at the end of the table.