Differentiate between target and source branch deletion (#220)
Currently, when either the source or the target branch of a pull request is deleted, the PR is rejected with either a generic "branch deleted" cause or, even worse, with a "source branch deleted" comment. This has no been fixed to correctly differentiate between target and/or source branch deletion. Uses a core fix from scm-manager/scm-manager#2141 in the newly introduced e2e tests.