Fix the generation of new PR IDs
The IDs for new pull requests have been created simply by&#10;adding 1 to the number of all existing pull requests for a&#10;repository. This works as long as new pull request has been&#10;deleted (however, for example by deleting the corresponding&#10;file from the storage directory). In this case, an existing&#10;number is generated overwriting the existing pull request.&#10;&#10;To prevent this, we now compute the maximum ID of all&#10;existing pull requests and add 1.

Pushed-by: Rene Pfeuffer<rene.pfeuffer@cloudogu.com>
Co-authored-by: René Pfeuffer<rene.pfeuffer@cloudogu.com>
Committed-by: René Pfeuffer<rene.pfeuffer@cloudogu.com>