Fix error with deleted source branch
In some circumstances it could happen, that the&#10;source branch of a pull request gets deleted but&#10;SCM-Manager does miss this event. If this branch&#10;has been used as the source branch for a pull&#10;request, this pull request could no longer&#10;be loaded, because a &#39;not found&#39; has been thrown.&#10;&#10;This catches this error and displays the pull request&#10;with a warning.

Co-authored-by: René Pfeuffer<rene.pfeuffer@cloudogu.com>