Fix performance issue for 'get all' request
To fetch a page on the overview page for pull requests, all pull
requests had been converted to dtos before they were partitioned for the
requested page. This could lead to performance issues, because for the
conversion all comments had to be read for each pull request.

With this commit, the conversion is done only for the pull requests
relevant for the selected page.