Fix Client-Error on SVN Repo Commit (#231)
This has been released with v3.11.1. This is just to close the
pull request.

Squash commits of branch feature/review_marks_in_file_tree:
- POC for review marks in file tree
- Migrate to wrapper
- Put check mark in front
- Use new extension
- Fix hover color
- Documentation

* Fix usage of BranchResolver on Unsupported Repo Types

This fixes a problem on SVN Repos that throws an error to the client on commits (because BranchResolver is not working on SVN Repos).

* Fix test shouldIgnorePushBecauseRepositoryDoesNotSupportPullRequests

fixes test so it more correctly working functions on SVN Repos

* rewrite test to fail on correctly on not supported repo

* Create fix_svn_post_commit_hook.yaml

---------

Co-authored-by: Rene Pfeuffer <rene.pfeuffer@cloudogu.com>